Introduce module coalesce
* Cut-paste Coalesce and derived adaptors into own module
* Run cargo fmt on new module
